+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## main
- 
+ - Added station metric units to constants.  
+
 ## 1.2.8 / 2023-08-21
  - Change resample method to use default Hanning window (was not using a window previously).
  - Define new Metrics class to improving handling of intensity metrics.